requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Redis
redis

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.16.0
shapely
geojson
requests

# Elasticsearch
elasticsearch==8.11.0

# Other utilities
python-dotenv==1.0.0
pytz==2025.2

# Storage
storageprovider-client==3.0.1

# Testing
pytest
httpx

oe-utils @ git+ssh://git@github.com/OnroerendErfgoed/oe_utils.git@7.2.0
oe-geoutils @ git+ssh://git@github.com/OnroerendErfgoed/oe_geoutils.git@3.0.0
storageprovider-client @ git+ssh://git@github.com/OnroerendErfgoed/storageprovider-client.git@3.0.1
oeauth @ git+ssh://git@github.com/OnroerendErfgoed/oeauth.git@2.10.0
